fix error output

+++ app/Http/Controllers/API/CarsController.php
@@ -11,7 +11,7 @@ use Illuminate\Support\Facades\Validator;
 class CarsController extends BaseController
 {
 
-    public function index()
+    public function index(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'limit' => 'numeric|min:2|max:100',
@@ -21,7 +21,6 @@ class CarsController extends BaseController
 
         $headers = [
             'X-Pagination-Total-Count' => $cars->total(),
-//            'X-Pagination-Page-Count' => $cars->lastPage(),
             'X-Pagination-Current-Page' => $cars->currentPage(),
             'X-Pagination-Per-Page' => $cars->perPage(),
         ];
@@ -35,7 +34,7 @@ class CarsController extends BaseController
         $car = Cars::find($id);
 
         if (is_null($car)) {
-            return response()->json('The car is not found!', 404);
+            return response()->json(['message' => 'The car is not found!'], 404);
         }
 
         return response()->json(new CarResource($car), 200);
@@ -47,7 +46,7 @@ class CarsController extends BaseController
             'title' => 'required|unique:cars|max:255',
         ]);
         if ($validator->fails()) {
-            return response()->json($validator->errors(), 422);
+            return response()->json(['message' => 'Validation error!', 'errors' => $validator->errors()], 422);
         }
 
         $car = new Cars();
@@ -61,13 +60,20 @@ class CarsController extends BaseController
         return response()->json(new CarResource($car), 201, $headers);
     }
 
-    public function update(Request $request, Cars $car)
+    public function update(Request $request, $id)
     {
+        /** @var Cars $car */
+        $car = Cars::find($id);
+
+        if (is_null($car)) {
+            return response()->json(['message' => 'The car is not found!'], 404);
+        }
+
         $validator = Validator::make($request->all(), [
             'title' => 'required|unique:cars|max:255',
         ]);
         if ($validator->fails()) {
-            return response()->json($validator->errors(), 422);
+            return response()->json(['message' => 'Validation error!', 'errors' => $validator->errors()], 422);
         }
 
         $car->fill($validator->validated());
@@ -76,12 +82,19 @@ class CarsController extends BaseController
         return response()->json(null, 204);
     }
 
-    public function destroy(Cars $car)
+    public function destroy($id)
     {
+        /** @var Cars $car */
+        $car = Cars::find($id);
+
+        if (is_null($car)) {
+            return response()->json(['message' => 'The car is not found!'], 404);
+        }
+
         try {
             $car->delete();
         } catch (\Exception $e) {
-            return response()->json($e->getMessage(), 500);
+            return response()->json(['message' => $e->getMessage()], 500);
         }
         return response()->json(null, 204);
     }
